created lscp model, example and test
